refactor: mentors pages
