+++ frontend/src/App.jsx
@@ -8,9 +8,9 @@ import Course from './pages/dashboard/courses/Course'
 import Lesson from './pages/dashboard/courses/Lesson'
 import Checkout from './pages/dashboard/Checkout'
 import Mentors from './pages/dashboard/mentors'
-import Mentor from './pages/dashboard/mentors/mentor'
-import MentorCourses from './pages/dashboard/mentors/mentor/options/MentorCourses'
-import Chat from './pages/dashboard/mentors/mentor/options/Chat'
+import Mentor from './pages/dashboard/mentors/Mentor'
+import MentorCourses from './pages/dashboard/mentors/MentorCourses'
+import Chat from './pages/dashboard/mentors/Chat'
 import TeacherHome from './pages/teacher/TeacherHome'
 import CoursesTeacher from './pages/teacher/CoursesTeacher'
 import NotFound from './pages/NotFound'
